Переносит в общие утилитарные функции метод из прошлого коммита, меняет humanizeDate, используя момент
