+++ src/utils/film.js
@@ -1,5 +1,3 @@
-import moment from 'moment';
-
 export const sortByDate = (filmA, filmB) => {
   return filmB.releaseDate.getTime() - filmA.releaseDate.getTime();
 };
@@ -10,12 +8,4 @@ export const sortByRating = (filmA, filmB) => {
 
 export const sortByComments = (filmA, filmB) => {
   return filmB.comments.length - filmA.comments.length;
-};
-
-export const getDuration = (timeInMinutes) => {
-  const duration = moment.duration(timeInMinutes, `m`);
-  const hours = duration.hours();
-  const minutes = duration.minutes();
-
-  return `${hours > 0 ? `${hours}h ` : ``}${minutes}m`;
 };
